+++ AAA_API/labels/get.php
@@ -18,14 +18,7 @@ else if(isset($userID)) {
 
 
 
-
-
-// If there's no result, return an error
-if($res === NULL) {
-	ApiResponse::httpResponse(404, ["error" => "The requested label could not be found."]);
-}
-
 // Properly return the results
-ApiResponse::httpResponse(200, [ "message" => "Label(s) found.", "data" => $res ]);
+ApiResponse::httpResponse(200, [ "message" => "Label(s) found.", "data" => ($res === NULL ? [] : $res) ]);
 
 ?>