+++ AAA_API/spotify/playlists.php
@@ -9,6 +9,12 @@ $data = SpotifyApi::getMyPlaylists();
 // Parse the data using the models
 $playlists = SpotifyCollection::createPlaylistCollection($data);
 
+// Get liked tracks
+@$liked->name = "Liked Songs";
+$liked->id = NULL;
+@$liked->numTracks = count(SpotifyApi::getMyLikedTracks());
+array_unshift($playlists->data, $liked);
+
 // Return the useful playlist data
 ApiResponse::httpResponse(200, ["message" => "Found your playlists.", "data" => $playlists->data]);